recommendation-api/requirements.txt
flask==2.0.1
werkzeug==2.0.3
pandas==1.3.0
numpy==1.21.0
torch==1.9.0
scikit-learn==0.24.2
transformers==4.9.1
mysql-connector-python==8.0.26
requests==2.26.0